fix(metrics-bridge): preserve partial peg counter events
